add tests for eth receipt.gasUsed
