+++ tests/evm_space/base.py
@@ -1,0 +1,67 @@
+#!/usr/bin/env python3
+import os, sys
+sys.path.insert(1, os.path.join(sys.path[0], '..'))
+
+from test_framework.test_framework import ConfluxTestFramework
+from conflux.rpc import RpcClient
+from web3 import Web3
+from test_framework.util import *
+from eth_utils import decode_hex
+from conflux.config import default_config
+
+class Web3Base(ConfluxTestFramework):
+    def set_test_params(self):
+        self.num_nodes = 1
+        self.conf_parameters["evm_chain_id"] = str(10)
+        self.conf_parameters["evm_transaction_block_ratio"] = str(1)
+
+    def setup_network(self):
+        self.add_nodes(self.num_nodes)
+        self.start_node(0, ["--archive"])
+        self.rpc = RpcClient(self.nodes[0])
+
+        ip = self.nodes[0].ip
+        port = self.nodes[0].ethrpcport
+        self.w3 = Web3(Web3.HTTPProvider(f'http://{ip}:{port}/'))
+        assert_equal(self.w3.isConnected(), True)
+
+    def cross_space_transfer(self, to, value):
+        to = to.replace('0x', '')
+
+        tx = self.rpc.new_tx(
+            value=value,
+            receiver="0x0888000000000000000000000000000000000006",
+            data=decode_hex(f"0xda8d5daf{to}000000000000000000000000"),
+            nonce=self.rpc.get_nonce(self.cfxAccount),
+            gas=1000000,
+        )
+
+        self.rpc.send_tx(tx, True)
+
+    def construct_evm_tx(self, receiver, data_hex, nonce):
+        signed = self.evmAccount.signTransaction({
+            "to": receiver,
+            "value": 0,
+            "gasPrice": 1,
+            "gas": 150000,
+            "nonce": nonce,
+            "chainId": 10,
+            "data": data_hex,
+        })
+
+        tx = [nonce, 1, 150000, bytes.fromhex(receiver.replace('0x', '')), 0, bytes.fromhex(data_hex.replace('0x', '')), signed["v"], signed["r"], signed["s"]]
+        return tx, signed["hash"]
+
+    def run_test(self):
+        self.cfxPrivkey = default_config['GENESIS_PRI_KEY']
+        self.cfxAccount = self.rpc.GENESIS_ADDR
+        print(f'Using Conflux account {self.cfxAccount}')
+        # initialize EVM account
+        self.evmAccount = self.w3.eth.account.privateKeyToAccount('0x0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef')
+        print(f'Using EVM account {self.evmAccount.address}')
+        self.cross_space_transfer(self.evmAccount.address, 1 * 10 ** 18)
+        assert_equal(self.nodes[0].eth_getBalance(self.evmAccount.address), hex(1 * 10 ** 18))
+
+
+if __name__ == "__main__":
+    EvmTx2ReceiptTest().main()
