+++ tests/evm_space/tx_and_receipt_test.py
@@ -1,39 +1,12 @@
 #!/usr/bin/env python3
-from test_framework.test_framework import ConfluxTestFramework
-from conflux.rpc import RpcClient
-from web3 import Web3
+import os, sys
+sys.path.insert(1, os.path.join(sys.path[0], '..'))
+
 from test_framework.util import *
-from eth_utils import decode_hex
 from conflux.config import default_config
+from base import Web3Base
 
-class EvmTx2ReceiptTest(ConfluxTestFramework):
-    def set_test_params(self):
-        self.num_nodes = 1
-        self.conf_parameters["evm_chain_id"] = str(10)
-        self.conf_parameters["evm_transaction_block_ratio"] = str(1)
-
-    def setup_network(self):
-        self.add_nodes(self.num_nodes)
-        self.start_node(0, ["--archive"])
-        self.rpc = RpcClient(self.nodes[0])
-
-        ip = self.nodes[0].ip
-        port = self.nodes[0].ethrpcport
-        self.w3 = Web3(Web3.HTTPProvider(f'http://{ip}:{port}/'))
-        assert_equal(self.w3.isConnected(), True)
-
-    def cross_space_transfer(self, to, value):
-        to = to.replace('0x', '')
-
-        tx = self.rpc.new_tx(
-            value=value,
-            receiver="0x0888000000000000000000000000000000000006",
-            data=decode_hex(f"0xda8d5daf{to}000000000000000000000000"),
-            nonce=self.rpc.get_nonce(self.cfxAccount),
-            gas=1000000,
-        )
-
-        self.rpc.send_tx(tx, True)
+class EvmTx2ReceiptTest(Web3Base):
 
     def run_test(self):
         self.cfxPrivkey = default_config['GENESIS_PRI_KEY']
@@ -43,7 +16,28 @@ class EvmTx2ReceiptTest(ConfluxTestFramework):
         self.evmAccount = self.w3.eth.account.privateKeyToAccount('0x0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef')
         print(f'Using EVM account {self.evmAccount.address}')
         self.cross_space_transfer(self.evmAccount.address, 1 * 10 ** 18)
-        assert_equal(self.nodes[0].ethrpc.eth_getBalance(self.evmAccount.address), hex(1 * 10 ** 18))
+        assert_equal(self.nodes[0].eth_getBalance(self.evmAccount.address), hex(1 * 10 ** 18))
+
+        nonce = self.w3.eth.getTransactionCount(self.evmAccount.address)
+        signed = self.evmAccount.signTransaction({
+            "to": self.evmAccount.address,
+            "value": 1,
+            "gasPrice": 1,
+            "gas": 210000,
+            "nonce": nonce,
+            "chainId": 10,
+        })
+
+        return_tx_hash = self.w3.eth.sendRawTransaction(signed["rawTransaction"])
+        self.rpc.generate_block(1)
+        self.rpc.generate_blocks(20, 1)
+        receipt = self.w3.eth.waitForTransactionReceipt(return_tx_hash)
+        assert_equal(receipt["status"], 1)
+        assert_equal(receipt["gasUsed"], 210000 / 4 * 3)
+        assert_equal(receipt["txExecErrorMsg"], None)
+
+        tx = self.w3.eth.get_transaction(return_tx_hash)
+        assert_equal(receipt["status"], 1)
 
 
 if __name__ == "__main__":
